fix eslint-warnings https://github.com/bluewings/gatsby-starter-mdx-blog/issues/11
