+++ .eslintrc.yml
@@ -1,7 +1,13 @@
 ---
 extends:
-  - airbnb-base
+  - airbnb
   - prettier
 env:
   browser: true
   es6: true
+rules:
+  import/prefer-default-export: 0
+  no-use-before-define: 0
+  react/forbid-prop-types: 0
+  react/jsx-one-expression-per-line: 0
+  react/prop-types: 0
