+++ package.json
@@ -30,6 +30,7 @@
     "parse-numeric-range": "0.0.2",
     "prism-react-renderer": "^0.1.5",
     "prism-themes": "^1.0.1",
+    "prop-types": "^15.7.1",
     "react": "^16.8.1",
     "react-dom": "^16.8.1",
     "react-helmet": "^5.2.0",
@@ -59,9 +60,11 @@
   },
   "devDependencies": {
     "eslint": "^5.3.0",
-    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^4.0.0",
     "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.12.4",
     "prettier": "1.16.4"
   }
 }
