+++ src/components/bio.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { StaticQuery, graphql } from 'gatsby';
 import Image from 'gatsby-image';
-
 import { rhythm } from '../utils/typography';
 
 function Bio() {
